update to latest version
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@41573 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/libnl-tiny/genl_mngt.c
@@ -154,7 +154,7 @@ int genl_register(struct nl_cache_ops *ops)
 		goto errout;
 	}
 
-	if (ops->co_hdrsize < GENL_HDRSIZE(0)) {
+	if ((size_t) ops->co_hdrsize < GENL_HDRSIZE(0)) {
 		err = -NLE_INVAL;
 		goto errout;
 	}
